requirements.txt
langchain
langchain-openai
langchain-community
python-dotenv
langgraph
ipykernel
tavily-python